pyproject.toml
[project]
name = "lysergic-dream-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "absl-py==2.3.1",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyascii==0.3.3",
    "attrs==25.4.0",
    "audioread==3.1.0",
    "babel==2.17.0",
    "bangla==0.0.5",
    "blinker==1.9.0",
    "blis==1.3.3",
    "bnnumerizer==0.0.2",
    "bnunicodenormalizer==0.1.7",
    "cachetools==6.2.4",
    "catalogue==2.0.10",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.1.8",
    "cloudpathlib==0.23.0",
    "confection==0.1.5",
    "contourpy==1.3.3",
    "coqpit==0.0.17",
    "cycler==0.12.1",
    "cymem==2.0.13",
    "cython==3.2.3",
    "dateparser==1.1.8",
    "decorator==5.2.1",
    "docopt==0.6.2",
    "einops==0.8.1",
    "encodec==0.1.1",
    "ffmpeg>=1.4",
    "filelock==3.20.1",
    "flask==3.1.2",
    "fonttools==4.61.1",
    "frozenlist==1.8.0",
    "fsspec==2025.12.0",
    "g2pkk==0.1.2",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.28.1",
    "google-api-python-client==2.187.0",
    "google-auth==2.45.0",
    "google-auth-httplib2==0.3.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.72.0",
    "grpcio==1.76.0",
    "grpcio-status==1.71.2",
    "hangul-romanize==0.1.0",
    "hf-xet==1.2.0",
    "httplib2==0.31.0",
    "huggingface-hub==0.36.0",
    "idna==3.11",
    "imageio==2.37.2",
    "imageio-ffmpeg==0.6.0",
    "inflect==7.5.0",
    "itsdangerous==2.2.0",
    "jamo==0.4.1",
    "jieba==0.42.1",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "jsonlines==1.2.0",
    "kiwisolver==1.4.9",
    "kokoro-onnx>=0.3.3",
    "lazy-loader==0.4",
    "librosa==0.11.0",
    "llvmlite==0.46.0",
    "markdown==3.10",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "more-itertools==10.8.0",
    "moviepy<2.0.0",
    "mpmath==1.3.0",
    "msgpack==1.1.2",
    "multidict==6.7.0",
    "murmurhash==1.0.15",
    "networkx==2.8.8",
    "nltk==3.9.2",
    "num2words==0.5.14",
    "numba==0.63.1",
    "numpy>=2.0.2",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.5",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvshmem-cu12==3.3.20",
    "nvidia-nvtx-cu12==12.8.90",
    "packaging==25.0",
    "pandas==1.5.3",
    "pillow==11.3.0",
    "platformdirs==4.5.1",
    "pooch==1.8.2",
    "preshed==3.0.12",
    "proglog==0.1.12",
    "propcache==0.4.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psutil==7.1.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pynndescent==0.5.13",
    "pyparsing==3.2.5",
    "pypinyin==0.55.0",
    "pysbd==0.3.4",
    "python-crfsuite==0.9.11",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "regex==2025.11.3",
    "requests==2.32.5",
    "rsa==4.9.1",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.16.3",
    "six==1.17.0",
    "smart-open==7.5.0",
    "soundfile==0.13.1",
    "soxr==1.0.0",
    "spacy==3.8.11",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.5.2",
    "sudachidict-core==20251022",
    "sudachipy==0.6.10",
    "sympy==1.14.0",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "thinc==8.3.10",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.1",
    "torch==2.9.1",
    "torchaudio==2.9.1",
    "tqdm==4.67.1",
    "trainer==0.0.36",
    "transformers==4.57.3",
    "typeguard==4.4.4",
    "typer-slim==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzlocal==5.3.1",
    "umap-learn==0.5.9.post2",
    "unidecode==1.4.0",
    "uritemplate==4.2.0",
    "urllib3==2.6.2",
    "wasabi==1.1.3",
    "weasel==0.4.3",
    "werkzeug==3.1.4",
    "wrapt==2.0.1",
    "yarl==1.22.0",
    "torch>=2.4.0",
    "torchaudio>=2.4.0",
    "onnxruntime>=1.19.0",
    "kokoro-onnx>=0.4.9",
    "soundfile",
    "requests",
    "python-dotenv",
    "google-api-python-client",
    "google-auth-oauthlib",
    "onnx>=1.20.0",
]
